+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [0.0.11]
 
 - Attach notes as context when create marp slide
+- Smart coding support
 
 ## [0.0.10]